Moduls

+++ Library/src/main/java/com/shop/library/service/ShoppingCartService.java
@@ -3,6 +3,7 @@ package com.shop.library.service;
 import com.shop.library.model.Customer;
 import com.shop.library.model.Product;
 import com.shop.library.model.ShoppingCart;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ShoppingCartService {
@@ -16,7 +17,7 @@ public interface ShoppingCartService {
 
     ShoppingCart deleteItemFromCart(Product product, Customer customer);
 
-    Double totalPrice(ShoppingCart shoppingCart);
+    BigDecimal totalPrice(ShoppingCart shoppingCart);
 
     List<ShoppingCart> findAll();
 }
